Add Qt leak-checking functionality. Make sure widgets get deleted when their dock is closed

+++ src/nspyre/misc/concurrency.py
@@ -94,5 +94,7 @@ class ProcessRunner:
 
     def kill(self):
         """Kill the process."""
-        self.proc.terminate()
-        self.proc.join()
+        if self.proc:
+            self.proc.terminate()
+            self.proc.join()
+            self.proc = None
